move markers in their own component
