+++ src/components/LeafletMap/style.scss
@@ -179,39 +179,3 @@ path.camera-area {
     z-index: 1010;
   }
 }
-.leaflet-popup-content {
-  margin: 0;
-  .device-popup {
-    display: flex;
-    flex-direction: column;
-    h3 {
-      display: block;
-      background-color: #666;
-      color: #f6f6f6;
-    }
-    &-information {
-      padding-bottom: 0.5rem;
-      p {
-        margin: 0;
-        line-height: 1.4;
-      }
-      &-label {
-        padding: 0;
-        font-size: 14px;
-        font-weight: 400;
-        line-height: 16px;
-        color: #999;
-        border: none;
-        box-shadow: none;
-      }
-      &-text {
-        padding: 4px 0 0;
-        font-size: 16px;
-        line-height: 1;
-        color: #333;
-        border: none;
-        box-shadow: none;
-      }
-    }
-  }
-}
